Added Tests for Milestone 4
Finished tests to make them more thourough. Here’s what I changed…

1) In DesignParser, changed the name of SingleClassTester to run. DOES
THE SAME THING ONLY CHANGED THE NAME. It now works in conjunction with
the constructor that passes in its class name (i.e public
DesignParser(String classname)

2) Created a much more thorough test process for the singleton class.
What it does is creates a designParser with a given className. From
there it looks at every line and checks for the “blue” string, which
signals if it is a singleton or not. If it pops up the test returns
true and passes, otherwise it returns false. If you want the specifics
on how it works please look at the comments on the first test
(interface1)